+++ deepof/data.py
@@ -85,7 +85,7 @@ from sklearn.preprocessing import (
 from tqdm import tqdm
 
 import deepof.annotation_utils
-from deepof.config import PROGRESS_BAR_FIXED_WIDTH, ROI_COLORS, suppress_warnings_context, DistanceUnit
+from deepof.config import PROGRESS_BAR_FIXED_WIDTH, ROI_COLORS, suppress_warnings_context, DistanceUnit, BitPrecision
 #import deepof.clustering.models_new
 import deepof.clustering.training
 import deepof.utils
@@ -132,7 +132,8 @@ def load_project(
         video_format: str = ".mp4",
         video_scale: int = 1,
         number_of_rois = 0,
-        fast_implementations_threshold: int = 50000,) -> coordinates:  # pragma: no cover
+        fast_implementations_threshold: int = 50000,
+        bit_precision: int = None,) -> coordinates:  # pragma: no cover
     """Load a pre-saved pickled Coordinates object. Will update Coordinate objects from older versions of deepof (down to 0.7) to work with this version. 
     Very old projects will be recreated during loading with the current version of Deepof. For this purpose input arguments can be set just as in a recular project definition.
 
@@ -162,6 +163,7 @@ def load_project(
         video_scale (int): diameter of the arena in mm (if the arena is round) or length of the first specified arena side (if the arena is polygonal).            
         number_of_rois (int): number of behavior rois to be drawn during project creation, default = 0,
         fast_implementations_threshold (int): If the total number of frames in the project is larger than this, numba implementations of all functions with a numba option will be used.
+        bit_precision (int): Minimum bit precision used for the biggest tables. Default is None for auto precision (64 and gets decreased for large projects). Can be manually set to 32 or 64
 
     Returns:
         coordinates (deepof_coordinates): Pre-run coordinates object.
@@ -174,6 +176,8 @@ def load_project(
 
     coordinates._project_path = os.path.split(project_path[0:-1])[0]
     # Small fix for compatibility of pre 0.9 versions
+    if not (hasattr(coordinates, "_bit_precision")):
+        coordinates._bit_precision = BitPrecision.parse(64)
     if not (hasattr(coordinates, "_custom_behaviors")):
             coordinates._custom_behaviors = None
             coordinates._custom_continuous_behavior_names=[]
@@ -232,12 +236,12 @@ def load_project(
             video_scale=str(coordinates._arena_dims)+ " mm",
             number_of_rois = number_of_rois,
             fast_implementations_threshold=fast_implementations_threshold,
+            bit_precision=bit_precision
         )
 
         coordinates=redone_project.create(force=True)
 
     return coordinates
-
 
 class Project:
     """Class for loading and preprocessing motion tracking data of individual and multiple animals.
@@ -273,6 +277,7 @@ class Project:
         number_of_rois: int = 0,
         frame_rate: float = None,
         fast_implementations_threshold: int = 50000,
+        bit_precision: int = None,
     ):
         """Initialize a Project object.
 
@@ -301,7 +306,7 @@ class Project:
             video_scale (int): diameter of the arena in mm (if the arena is round) or length of the first specified arena side (if the arena is polygonal).            
             number_of_rois (int): number of behavior rois to be drawn during project creation, default = 0,
             fast_implementations_threshold (int): If the total number of frames in the project is larger than this, numba implementations of all functions with a numba option will be used.
-
+            bit_precision (int): Minimum bit precision used for the biggest tables. Default is None for auto precision (64 and gets decreased for large projects). Can be manually set to 32 or 64
         """
         # Set version
         self.version=current_deepof_version
@@ -313,6 +318,7 @@ class Project:
         self.table_path = table_path
         self.source_table_path = table_path
         self.trained_path = os.path.join(project_path, project_name, "trained_models")
+
 
         # Detect files to load from disk
         self.table_format = table_format
@@ -435,33 +441,20 @@ class Project:
             self.run_numba = True
         if frames_max > 360000 or frames_sum > 900000: #roughly one 4 hour video at 25 fps or 10 hours of recording material in total
             self.very_large_project = True
+        if bit_precision is None:
+            self.bit_precision = BitPrecision.parse(64)
+            if self.very_large_project:
+                self.bit_precision = BitPrecision.parse(32)
+                print(
+                    f"\033[33mInfo! Large project detected! Reduced bit precision to 32 bit! (You can manually set it to 32, or 64 bit with input bit_precision.)\033[0m"
+                )
+        else:
+            self.bit_precision = BitPrecision.parse(bit_precision)
+
 
         # If the bodypart names in his table deviate from the ones deepOF expects, the user can rename them 
-        rename_bodyparts_dict = None
-        if rename_bodyparts is not None and isinstance(rename_bodyparts, list) and "npy" not in table_format:
-            if len(rename_bodyparts) == 8:
-                pattern=deepof.utils.connect_mouse(animal_ids="", graph_preset="deepof_8").nodes                            
-            elif len(rename_bodyparts) == 11:
-                pattern=deepof.utils.connect_mouse(animal_ids="", graph_preset="deepof_11").nodes            
-            elif len(rename_bodyparts) == 14:
-                pattern=deepof.utils.connect_mouse(animal_ids="", graph_preset="deepof_14").nodes            
-            else: # pragma: no cover
-                raise NotImplementedError(f"Number of custom bodypart names should be 8, 11 or 14 but your list has {len(rename_bodyparts)} elements!") 
-            
-            # Creates a dictionary assigning table bp names to corresponding deepOF bp names
-            rename_bodyparts_dict = {}
-            node_assignment_string = f"Your custom bodypart names will be assigned to deepOF bodyparts as follows:\n"
-            for deepof_node, custom_node in zip(pattern, rename_bodyparts):
-                node_assignment_string += f"{deepof_node} : {custom_node},\n"
-                rename_bodyparts_dict[custom_node] = deepof_node
-            node_assignment_string+=f"If this assignment is incorrect, please update your \"bodypart_names\" list with the correct order and rerun the project creation" 
-            print(node_assignment_string)
-        # Special case table format npy with no headers. Here rename_bodyparts functions as a list of bodypart names with only the keys being used
-        elif "npy" in table_format:
-            rename_bodyparts_dict={}
-            for bp in rename_bodyparts:
-                rename_bodyparts_dict[bp]=bp
-
+        rename_bodyparts_dict = self.rename_bodyparts(rename_bodyparts, table_format)
+        
 
         # Init the rest of the parameters
         self.angles = True
@@ -553,6 +546,34 @@ class Project:
                 "Project already exists. Delete it or specify a different name."
             )  # pragma: no cover
 
+
+    def rename_bodyparts(self, rename_bodyparts, table_format):
+        rename_bodyparts_dict = None
+        if rename_bodyparts is not None and isinstance(rename_bodyparts, list) and "npy" not in table_format:
+            if len(rename_bodyparts) == 8:
+                pattern=deepof.utils.connect_mouse(animal_ids="", graph_preset="deepof_8").nodes                            
+            elif len(rename_bodyparts) == 11:
+                pattern=deepof.utils.connect_mouse(animal_ids="", graph_preset="deepof_11").nodes            
+            elif len(rename_bodyparts) == 14:
+                pattern=deepof.utils.connect_mouse(animal_ids="", graph_preset="deepof_14").nodes            
+            else: # pragma: no cover
+                raise NotImplementedError(f"Number of custom bodypart names should be 8, 11 or 14 but your list has {len(rename_bodyparts)} elements!") 
+            
+            # Creates a dictionary assigning table bp names to corresponding deepOF bp names
+            rename_bodyparts_dict = {}
+            node_assignment_string = f"Your custom bodypart names will be assigned to deepOF bodyparts as follows:\n"
+            for deepof_node, custom_node in zip(pattern, rename_bodyparts):
+                node_assignment_string += f"{deepof_node} : {custom_node},\n"
+                rename_bodyparts_dict[custom_node] = deepof_node
+            node_assignment_string+=f"If this assignment is incorrect, please update your \"bodypart_names\" list with the correct order and rerun the project creation" 
+            print(node_assignment_string)
+        # Special case table format npy with no headers. Here rename_bodyparts functions as a list of bodypart names with only the keys being used
+        elif "npy" in table_format:
+            rename_bodyparts_dict={}
+            for bp in rename_bodyparts:
+                rename_bodyparts_dict[bp]=bp
+        
+        return rename_bodyparts_dict
 
     def load_start_markers(self, filepath): # pragma: no cover
         """Load start markers analogous to experimental conditions and do some checks"""
@@ -971,8 +992,8 @@ class Project:
                 quality_path = os.path.join(save_dir, f"{key}_likelihood")
                 table_path = os.path.join(save_dir, key)
                 
-                final_lik_dict[key] = save_dt(lik_dict_single[key], quality_path, self.very_large_project)
-                final_tab_dict[key] = save_dt(table_dict_single[key], table_path, self.very_large_project)
+                final_lik_dict[key] = save_dt(lik_dict_single[key].astype(self.bit_precision.dtype), quality_path, self.very_large_project)
+                final_tab_dict[key] = save_dt(table_dict_single[key].astype(self.bit_precision.dtype), table_path, self.very_large_project)
                 
                 pbar.update(1)
 
@@ -1019,7 +1040,7 @@ class Project:
 
                 #save scaled table
                 distance_path = os.path.join(self.project_path, self.project_name, 'Tables',key, key)
-                tab_dict[key] = save_dt(tab,distance_path,self.very_large_project)
+                tab_dict[key] = save_dt(tab.astype(self.bit_precision.dtype),distance_path,self.very_large_project)
 
                 pbar.update()
 
@@ -1054,7 +1075,7 @@ class Project:
 
                 #save distances for active table
                 distance_path = os.path.join(self.project_path, self.project_name, 'Tables',key, key + '_dist')
-                distance_dict[key] = save_dt(distance_tab,distance_path,self.very_large_project)
+                distance_dict[key] = save_dt(distance_tab.astype(self.bit_precision.dtype),distance_path,self.very_large_project)
 
                 #clean up
                 del distance_tab
@@ -1082,7 +1103,7 @@ class Project:
             i in tab.columns.levels[0] for i in nodes
         ], "Nodes should correspond to existent bodyparts"
 
-        distance_tab = deepof.utils.bpart_distance(tab)
+        distance_tab = deepof.utils.bpart_distance(tab, self.bit_precision)
         distance_tab = distance_tab.loc[
                 :, [np.all([i in nodes for i in j]) for j in distance_tab.columns]
             ]
@@ -1144,7 +1165,7 @@ class Project:
 
                     # get path for saving
                     angle_path = os.path.join(self.project_path, self.project_name, 'Tables',key, key + '_angle')
-                    angle_dict[key] = save_dt(dats,angle_path,self.very_large_project)
+                    angle_dict[key] = save_dt(dats.astype(self.bit_precision.dtype),angle_path,self.very_large_project)
                     pbar.update()
 
 
@@ -1259,7 +1280,7 @@ class Project:
 
                     # change dictionary to table and check size
                     areas_table = pd.DataFrame(
-                        areas_animal_dict, index=current_animal_table.index
+                        areas_animal_dict, index=current_animal_table.index, dtype=self.bit_precision.dtype
                     )
                     if animal_id is not None:
                         areas_table.columns = [
@@ -1273,7 +1294,7 @@ class Project:
                     current_table = pd.concat([current_table, areas_table], axis=1)
 
                 area_path = os.path.join(self.project_path, self.project_name, 'Tables',key, key + '_area')
-                all_areas_dict[key] = save_dt(current_table,area_path,self.very_large_project)
+                all_areas_dict[key] = save_dt(current_table.astype(self.bit_precision.dtype),area_path,self.very_large_project)
                 pbar.update()
 
         if not_all_area_warn:
@@ -1431,6 +1452,7 @@ class Project:
             run_numba=self.run_numba,
             very_large_project=self.very_large_project,
             version=self.version,
+            bit_precision=self.bit_precision,
         )
 
         #set supervised parameters via initial reset (sets values to defaults)
@@ -1597,6 +1619,7 @@ class Coordinates:
         run_numba: bool = False,
         very_large_project: bool = False,
         version: str = None,
+        bit_precision: int = 64,
     ):
         """Class for storing the results of a ran project. Methods are mostly setters and getters in charge of tidying up the generated tables.
 
@@ -1628,6 +1651,7 @@ class Coordinates:
             run_numba (bool): Determines if numba versions of functions should be used (run faster but require initial compilation time on first run)
             very_large_project (bool): Decides if memory efficient data loading and saving should be used
             version (str): version of deepof this object was created with
+            bit_precision (int): Minimum bit precision used for the biggest tables. Default is 0 for auto precision (64 and gets decreased for large projects). Can be manually set to 16, 32 or 64
 
         """
         self._project_path = project_path
@@ -1660,6 +1684,7 @@ class Coordinates:
         self._run_numba = run_numba
         self._very_large_project = very_large_project
         self._version = version
+        self._bit_precision = bit_precision
         self._custom_behaviors = None
         self._custom_continuous_behavior_names=[]
 
@@ -2018,7 +2043,7 @@ class Coordinates:
         # 10. Handle missing animals based on quality data
         table_dict = deepof.utils.set_missing_animals(self, {key: tab}, quality)
         
-        return table_dict[key]
+        return table_dict[key].astype(self._bit_precision.dtype)
            
             
     def get_distances(
@@ -2140,7 +2165,7 @@ class Coordinates:
             
             tab = tab.loc[:, list(set(sorted_edges) & set(tab.columns))]
       
-        return tab
+        return tab.astype(self._bit_precision.dtype)
     
    
     def get_angles(
@@ -2257,7 +2282,7 @@ class Coordinates:
         # 6. Handle missing animals based on quality data
         tab = deepof.utils.set_missing_animals(self, {key: tab}, quality)[key]
 
-        return tab
+        return tab.astype(self._bit_precision.dtype)
     
   
     def get_areas(
@@ -2370,7 +2395,7 @@ class Coordinates:
         # 6. Handle missing animals based on quality data
         tab = deepof.utils.set_missing_animals(self, {key: tab}, quality)[key]
     
-        return tab
+        return tab.astype(self._bit_precision.dtype)
 
 
     def get_videos(self, full_paths: bool = False, play: bool = False):
@@ -2872,7 +2897,7 @@ class Coordinates:
                         #load table if not already loaded
                         result = get_dt(to_preprocess[k], key, return_path=True)
                         if result and len(result)==2 :
-                            tab = result[0]
+                            tab = result[0].astype(self._bit_precision.dtype)
                             table_path = result[1]
                         tab_nodes = tab[:, :, node_sorting_indices]
                         tab_edges = tab[:, :, edge_sorting_indices] 
@@ -2920,7 +2945,7 @@ class Coordinates:
                         quality = get_dt(quality_to_load, key)
                         tab[speed_feature_names] = quality[speed_feature_names]
 
-                    tab = np.array(tab)
+                    tab = np.array(tab).astype(self._bit_precision.dtype)
 
                     # Split node features (positions, speeds) from edge features (distances)
                     dataset = (